+++ src/bl64-api-env.bash
@@ -2,17 +2,17 @@
 # BashLib64 / Module / Globals / Interact with RESTful APIs
 #######################################
 
-export BL64_API_VERSION='1.1.1'
+declare BL64_API_VERSION='2.0.0'
 
-export BL64_API_MODULE='0'
+declare BL64_API_MODULE='0'
 
 #
 # Common constants
 #
 
-export BL64_API_METHOD_DELETE='DELETE'
-export BL64_API_METHOD_GET='GET'
-export BL64_API_METHOD_POST='POST'
-export BL64_API_METHOD_PUT='PUT'
+declare BL64_API_METHOD_DELETE='DELETE'
+declare BL64_API_METHOD_GET='GET'
+declare BL64_API_METHOD_POST='POST'
+declare BL64_API_METHOD_PUT='PUT'
 
-export _BL64_API_TXT_ERROR_API_FAILED='API call failed'
+declare _BL64_API_TXT_ERROR_API_FAILED='API call failed'