+++ src/bl64-arc-env.bash
@@ -2,16 +2,16 @@
 # BashLib64 / Module / Globals / Manage archive files
 #######################################
 
-export BL64_ARC_VERSION='2.2.2'
+declare BL64_ARC_VERSION='3.0.0'
 
-export BL64_ARC_MODULE='0'
+declare BL64_ARC_MODULE='0'
 
-export BL64_ARC_CMD_TAR=''
-export BL64_ARC_CMD_UNZIP=''
+declare BL64_ARC_CMD_TAR=''
+declare BL64_ARC_CMD_UNZIP=''
 
-export BL64_ARC_SET_TAR_VERBOSE=''
+declare BL64_ARC_SET_TAR_VERBOSE=''
 
-export BL64_ARC_SET_UNZIP_OVERWRITE=''
+declare BL64_ARC_SET_UNZIP_OVERWRITE=''
 
-export _BL64_ARC_TXT_OPEN_ZIP='open zip archive'
-export _BL64_ARC_TXT_OPEN_TAR='open tar archive'
+declare _BL64_ARC_TXT_OPEN_ZIP='open zip archive'
+declare _BL64_ARC_TXT_OPEN_TAR='open tar archive'