+++ src/bl64-bsh-env.bash
@@ -2,10 +2,10 @@
 # BashLib64 / Module / Globals / Interact with Bash shell
 #######################################
 
-export BL64_BSH_VERSION='2.1.0'
+declare BL64_BSH_VERSION='3.0.0'
 
-export BL64_BSH_MODULE='0'
+declare BL64_BSH_MODULE='0'
 
-export BL64_BSH_VERSION_BASH=''
+declare BL64_BSH_VERSION_BASH=''
 
-export _BL64_BSH_TXT_IMPORT_YAML='convert and import shell variables from YAML file'
+declare _BL64_BSH_TXT_IMPORT_YAML='convert and import shell variables from YAML file'