+++ src/bl64-hlm-env.bash
@@ -2,20 +2,19 @@
 # BashLib64 / Module / Globals / Interact with HLM
 #######################################
 
-export BL64_HLM_VERSION='1.3.1'
+declare BL64_HLM_VERSION='2.0.0'
 
-# Optional module. Not enabled by default
-export BL64_HLM_MODULE='0'
+declare BL64_HLM_MODULE='0'
 
-export BL64_HLM_CMD_HELM="$BL64_VAR_UNAVAILABLE"
+declare BL64_HLM_CMD_HELM="$BL64_VAR_UNAVAILABLE"
 
-export BL64_HLM_SET_DEBUG=''
-export BL64_HLM_SET_OUTPUT_TABLE=''
-export BL64_HLM_SET_OUTPUT_JSON=''
-export BL64_HLM_SET_OUTPUT_YAML=''
+declare BL64_HLM_SET_DEBUG=''
+declare BL64_HLM_SET_OUTPUT_TABLE=''
+declare BL64_HLM_SET_OUTPUT_JSON=''
+declare BL64_HLM_SET_OUTPUT_YAML=''
 
-export BL64_HLM_RUN_TIMEOUT=''
+declare BL64_HLM_RUN_TIMEOUT=''
 
-export BL64_HLM_TXT_ADD_REPO='add Helm repository'
-export BL64_HLM_TXT_UPDATE_REPO='update Helm repository catalog'
-export BL64_HLM_TXT_DEPLOY_CHART='deploy helm chart'
+declare BL64_HLM_TXT_ADD_REPO='add Helm repository'
+declare BL64_HLM_TXT_UPDATE_REPO='update Helm repository catalog'
+declare BL64_HLM_TXT_DEPLOY_CHART='deploy helm chart'