+++ src/bl64-rnd-env.bash
@@ -2,27 +2,27 @@
 # BashLib64 / Module / Globals / Generate random data
 #######################################
 
-export BL64_RND_VERSION='1.2.1'
+declare BL64_RND_VERSION='2.0.0'
 
-export BL64_RND_MODULE='0'
+declare BL64_RND_MODULE='0'
 
-declare -ig BL64_RND_LENGTH_1=1
-declare -ig BL64_RND_LENGTH_20=20
-declare -ig BL64_RND_LENGTH_100=100
-declare -ig BL64_RND_RANDOM_MIN=0
-declare -ig BL64_RND_RANDOM_MAX=32767
+declare -i BL64_RND_LENGTH_1=1
+declare -i BL64_RND_LENGTH_20=20
+declare -i BL64_RND_LENGTH_100=100
+declare -i BL64_RND_RANDOM_MIN=0
+declare -i BL64_RND_RANDOM_MAX=32767
 
 # shellcheck disable=SC2155
-export BL64_RND_POOL_UPPERCASE="$(printf '%b' "$(printf '\\%o' {65..90})")"
-export BL64_RND_POOL_UPPERCASE_MAX_IDX="$(( ${#BL64_RND_POOL_UPPERCASE} - 1 ))"
+declare BL64_RND_POOL_UPPERCASE="$(printf '%b' "$(printf '\\%o' {65..90})")"
+declare BL64_RND_POOL_UPPERCASE_MAX_IDX="$(( ${#BL64_RND_POOL_UPPERCASE} - 1 ))"
 # shellcheck disable=SC2155
-export BL64_RND_POOL_LOWERCASE="$(printf '%b' "$(printf '\\%o' {97..122})")"
-export BL64_RND_POOL_LOWERCASE_MAX_IDX="$(( ${#BL64_RND_POOL_LOWERCASE} - 1 ))"
+declare BL64_RND_POOL_LOWERCASE="$(printf '%b' "$(printf '\\%o' {97..122})")"
+declare BL64_RND_POOL_LOWERCASE_MAX_IDX="$(( ${#BL64_RND_POOL_LOWERCASE} - 1 ))"
 # shellcheck disable=SC2155
-export BL64_RND_POOL_DIGITS="$(printf '%b' "$(printf '\\%o' {48..57})")"
-export BL64_RND_POOL_DIGITS_MAX_IDX="$(( ${#BL64_RND_POOL_DIGITS} - 1 ))"
-export BL64_RND_POOL_ALPHANUMERIC="${BL64_RND_POOL_UPPERCASE}${BL64_RND_POOL_LOWERCASE}${BL64_RND_POOL_DIGITS}"
-export BL64_RND_POOL_ALPHANUMERIC_MAX_IDX="$(( ${#BL64_RND_POOL_ALPHANUMERIC} - 1 ))"
+declare BL64_RND_POOL_DIGITS="$(printf '%b' "$(printf '\\%o' {48..57})")"
+declare BL64_RND_POOL_DIGITS_MAX_IDX="$(( ${#BL64_RND_POOL_DIGITS} - 1 ))"
+declare BL64_RND_POOL_ALPHANUMERIC="${BL64_RND_POOL_UPPERCASE}${BL64_RND_POOL_LOWERCASE}${BL64_RND_POOL_DIGITS}"
+declare BL64_RND_POOL_ALPHANUMERIC_MAX_IDX="$(( ${#BL64_RND_POOL_ALPHANUMERIC} - 1 ))"
 
-export _BL64_RND_TXT_LENGHT_MIN='length can not be less than'
-export _BL64_RND_TXT_LENGHT_MAX='length can not be greater than'
+declare _BL64_RND_TXT_LENGHT_MIN='length can not be less than'
+declare _BL64_RND_TXT_LENGHT_MAX='length can not be greater than'