+++ src/bl64-ui-env.bash
@@ -2,12 +2,12 @@
 # BashLib64 / Module / Globals / User Interface
 #######################################
 
-export BL64_UI_VERSION='1.1.1'
+declare BL64_UI_VERSION='2.0.0'
 
-export BL64_UI_MODULE='0'
+declare BL64_UI_MODULE='0'
 
-export BL64_UI_READ_TIMEOUT='60'
+declare BL64_UI_READ_TIMEOUT='60'
 
-export _BL64_UI_TXT_CONFIRMATION_QUESTION='Please confirm the operation by writting the message'
-export _BL64_UI_TXT_CONFIRMATION_MESSAGE='confirm-operation'
-export _BL64_UI_TXT_CONFIRMATION_ERROR='provided confirmation message is not what is expected'
+declare _BL64_UI_TXT_CONFIRMATION_QUESTION='Please confirm the operation by writting the message'
+declare _BL64_UI_TXT_CONFIRMATION_MESSAGE='confirm-operation'
+declare _BL64_UI_TXT_CONFIRMATION_ERROR='provided confirmation message is not what is expected'