+++ src/bl64-vcs-env.bash
@@ -2,26 +2,26 @@
 # BashLib64 / Module / Globals / Manage Version Control System
 #######################################
 
-export BL64_VCS_VERSION='1.15.2'
+declare BL64_VCS_VERSION='2.0.0'
 
-export BL64_VCS_MODULE='0'
+declare BL64_VCS_MODULE='0'
 
-export BL64_VCS_CMD_GIT=''
+declare BL64_VCS_CMD_GIT=''
 
-export BL64_VCS_SET_GIT_NO_PAGER=''
-export BL64_VCS_SET_GIT_QUIET=''
+declare BL64_VCS_SET_GIT_NO_PAGER=''
+declare BL64_VCS_SET_GIT_QUIET=''
 
 #
 # GitHub related parameters
 #
 
 # GitHub API FQDN
-export BL64_VCS_GITHUB_API_URL='https://api.github.com'
+declare BL64_VCS_GITHUB_API_URL='https://api.github.com'
 # Target GitHub public API version
-export BL64_VCS_GITHUB_API_VERSION='2022-11-28'
+declare BL64_VCS_GITHUB_API_VERSION='2022-11-28'
 # Special tag for latest release
-export BL64_VCS_GITHUB_LATEST='latest'
+declare BL64_VCS_GITHUB_LATEST='latest'
 
-export _BL64_VCS_TXT_CLONE_REPO='clone single branch from GIT repository'
-export _BL64_VCS_TXT_GET_LATEST_RELEASE='get release tag from latest release'
-export _BL64_VCS_TXT_GET_LATEST_RELEASE_FAILED='unable to determine latest release'
+declare _BL64_VCS_TXT_CLONE_REPO='clone single branch from GIT repository'
+declare _BL64_VCS_TXT_GET_LATEST_RELEASE='get release tag from latest release'
+declare _BL64_VCS_TXT_GET_LATEST_RELEASE_FAILED='unable to determine latest release'