+++ CHANGELOG.md
@@ -5,6 +5,25 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [19.0.0]
+
+### Added
+
+- cryp
+  - new module for cryptography tools: GPG, OpenSSL
+
+### Changed
+
+- all
+  - global variables are now declared by default instead of exported to enhance script re-entry (fork)
+- py
+  - bl64_py_venv_check: improved check method
+- rxtx
+  - bl64_rxtx_web_get_file: added param for setting owner, group
+- pkg
+  - **Breaking change**: added module dependency: `cryp`
+  - bl64_pkg_repository_add: added GPG key dearmor for Ubuntu
+
 ## [18.1.1]
 
 ### Added
@@ -573,6 +592,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - check
   - bl64_check_alert_failed
 
+[19.0.0]: https://github.com/automation64/bashlib64/compare/18.1.1...19.0.0
 [18.1.1]: https://github.com/automation64/bashlib64/compare/18.0.0...18.1.1
 [18.0.0]: https://github.com/automation64/bashlib64/compare/17.3.0...18.0.0
 [17.3.0]: https://github.com/automation64/bashlib64/compare/17.2.0...17.3.0